+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = hum
-version = 0.0.7
+version = 0.0.8
 url = https://github.com/otosense/hum
 platforms = any
 description_file = README.md
